+++ bindings/matlab/+nirs4all/capabilityManifest.m
@@ -5,6 +5,7 @@ manifest = struct();
 manifest.schema = 'nirs4all-core.capabilities.v1';
 manifest.aggregate = 'nirs4all-core';
 manifest.runtimeSurfaces = nirs4all.runtimeSurfaces();
+manifest.runtimeContracts = nirs4all.runtimeContracts();
 manifest.portableOperatorClasses = nirs4all.portableOperatorClasses();
 manifest.controllers = nirs4all.controllerCapabilities();
 end